Task 1 Uploaded
Draw activity diagrams (with swimlanes) for the following requirements:
Request for a change is submitted by the customer. It is first verified by the change manager. If it is not specified correctly, the customer is notified and the process ends. If it is correctly specified, the type of change is identified and then the request is sent for analysis to either the analyst, analyst and designer, or analyst, designer and manager of the development team. After the opinions of all the specialists are submitted, they are compared by the change manager and the change is either approved or rejected. The customer is notified about the decision and if the decision is to approve the change, the manager of the development team is also notified about it, then the process ends.